+++ wrapper/requests/guild_template.go
@@ -1,0 +1,56 @@
+package requests
+
+import "github.com/switchupcb/disgo/wrapper/resources"
+
+// Get Guild Template
+// GET /guilds/templates/{template.code}
+// https://discord.com/developers/docs/resources/guild-template#get-guild-template
+type GetGuildTemplate struct {
+	TemplateCode string `json:"code,omitempty"`
+}
+
+// Create Guild from Guild Template
+// POST /guilds/templates/{template.code}
+// https://discord.com/developers/docs/resources/guild-template#create-guild-from-guild-template
+type CreateGuildfromGuildTemplate struct {
+	Name string `json:"name,omitempty"`
+	Icon string `json:"icon,omitempty"`
+}
+
+// Get Guild Templates
+// GET /guilds/{guild.id}/templates
+// https://discord.com/developers/docs/resources/guild-template#get-guild-templates
+type GetGuildTemplates struct {
+	GuildID resources.Snowflake
+}
+
+// Create Guild Template
+// POST /guilds/{guild.id}/templates
+// https://discord.com/developers/docs/resources/guild-template#create-guild-template
+type CreateGuildTemplate struct {
+	Name        string  `json:"name,omitempty"`
+	Description *string `json:"description,omitempty"`
+}
+
+// Sync Guild Template
+// PUT /guilds/{guild.id}/templates/{template.code}
+// https://discord.com/developers/docs/resources/guild-template#sync-guild-template
+type SyncGuildTemplate struct {
+	TemplateCode string `json:"code,omitempty"`
+}
+
+// Modify Guild Template
+// PATCH /guilds/{guild.id}/templates/{template.code}
+// https://discord.com/developers/docs/resources/guild-template#modify-guild-template
+type ModifyGuildTemplate struct {
+	TemplateCode string  `json:"code,omitempty"`
+	Name         string  `json:"name,omitempty"`
+	Description  *string `json:"description,omitempty"`
+}
+
+// Delete Guild Template
+// DELETE /guilds/{guild.id}/templates/{template.code}
+// https://discord.com/developers/docs/resources/guild-template#delete-guild-template
+type DeleteGuildTemplate struct {
+	TemplateCode string `json:"code,omitempty"`
+}
