+++ wrapper/requests/oauth2.go
@@ -1,0 +1,15 @@
+package requests
+
+// Get Current Bot Application Information
+// GET /oauth2/applications/@me
+// https://discord.com/developers/docs/topics/oauth2#get-current-bot-application-information
+type GetCurrentBotApplicationInformation struct {
+	// TODO
+}
+
+// Get Current Authorization Information
+// GET /oauth2/@me
+// https://discord.com/developers/docs/topics/oauth2#get-current-authorization-information
+type GetCurrentAuthorizationInformation struct {
+	// TODO
+}
