+++ wrapper/requests/responses/threads.go
@@ -6,7 +6,7 @@ import (
 
 // List Active Threads Response Body
 // https://discord.com/developers/docs/resources/channel#list-active-threads-response-body
-type ListActiveThreads struct {
+type ListActiveThreadsResponse struct {
 	Threads []*resources.Channel      `json:"threads"`
 	Members []*resources.ThreadMember `json:"members"`
 	HasMore bool                      `json:"has_more"`
@@ -14,7 +14,7 @@ type ListActiveThreads struct {
 
 // List Public Archived Threads Response Body
 // https://discord.com/developers/docs/resources/channel#list-active-threads-response-body
-type ListPublicArchivedThreads struct {
+type ListPublicArchivedThreadsResponse struct {
 	Threads []*resources.Channel      `json:"threads"`
 	Members []*resources.ThreadMember `json:"members"`
 	HasMore bool                      `json:"has_more"`
@@ -22,7 +22,7 @@ type ListPublicArchivedThreads struct {
 
 // List Private Archived Threads Response Body
 // https://discord.com/developers/docs/resources/channel#list-active-threads-response-body
-type ListPrivateArchivedThreads struct {
+type ListPrivateArchivedThreadsResponse struct {
 	Threads []*resources.Channel      `json:"threads"`
 	Members []*resources.ThreadMember `json:"members"`
 	HasMore bool                      `json:"has_more"`
@@ -30,7 +30,7 @@ type ListPrivateArchivedThreads struct {
 
 // List Joined Private Archived Threads Response Body
 // https://discord.com/developers/docs/resources/channel#list-active-threads-response-body
-type ListJoinedPrivateArchivedThreads struct {
+type ListJoinedPrivateArchivedThreadsResponse struct {
 	Threads []*resources.Channel      `json:"threads"`
 	Members []*resources.ThreadMember `json:"members"`
 	HasMore bool                      `json:"has_more"`
