+++ wrapper/requests/sticker.go
@@ -1,0 +1,59 @@
+package requests
+
+import "github.com/switchupcb/disgo/wrapper/resources"
+
+// Get Sticker
+// GET /stickers/{sticker.id}
+// https://discord.com/developers/docs/resources/sticker#get-sticker
+type GetSticker struct {
+	StickerID resources.Snowflake
+}
+
+// List Nitro Sticker Packs
+// GET /sticker-packs
+// https://discord.com/developers/docs/resources/sticker#list-nitro-sticker-packs
+type ListNitroStickerPacks struct {
+	StickerPacks []*resources.StickerPack `json:"sticker_packs,omitempty"`
+}
+
+// List Guild Stickers
+// GET /guilds/{guild.id}/stickers
+// https://discord.com/developers/docs/resources/sticker#list-guild-stickers
+type ListGuildStickers struct {
+	GuildID resources.Snowflake
+}
+
+// Get Guild Sticker
+// GET /guilds/{guild.id}/stickers/{sticker.id}
+// https://discord.com/developers/docs/resources/sticker#get-guild-sticker
+type GetGuildSticker struct {
+	StickerID resources.Snowflake
+}
+
+// Create Guild Sticker
+// POST /guilds/{guild.id}/stickers
+// https://discord.com/developers/docs/resources/sticker#create-guild-sticker
+type CreateGuildSticker struct {
+	GuildID     resources.Snowflake
+	Name        string  `json:"name,omitempty"`
+	Description string  `json:"description,omitempty"`
+	Tags        *string `json:"tags,omitempty"`
+	Files       []byte  `disgo:"TODO"`
+}
+
+// Modify Guild Sticker
+// PATCH /guilds/{guild.id}/stickers/{sticker.id}
+// https://discord.com/developers/docs/resources/sticker#modify-guild-sticker
+type ModifyGuildSticker struct {
+	StickerID   resources.Snowflake
+	Name        string  `json:"name,omitempty"`
+	Description string  `json:"description,omitempty"`
+	Tags        *string `json:"tags,omitempty"`
+}
+
+// Delete Guild Sticker
+// DELETE /guilds/{guild.id}/stickers/{sticker.id}
+// https://discord.com/developers/docs/resources/sticker#delete-guild-sticker
+type DeleteGuildSticker struct {
+	StickerID resources.Snowflake
+}
