+++ wrapper/requests/voice.go
@@ -1,0 +1,8 @@
+package requests
+
+// List Voice Regions
+// GET /voice/regions
+// https://discord.com/developers/docs/resources/voice#list-voice-regions
+type ListVoiceRegions struct {
+	// TODO
+}
